+++ frontend/src/App.tsx
@@ -11,6 +11,8 @@ import { ChatPanel } from './panels/ChatPanel'
 import { ContextPanel } from './panels/ContextPanel'
 import { McpPanel } from './panels/McpPanel'
 import { MemoryPanel } from './panels/MemoryPanel'
+import { MemoryExtractPanel } from './panels/MemoryExtractPanel'
+import { MemoryComparePanel } from './panels/MemoryComparePanel'
 import { MultiAgentPanel } from './panels/MultiAgentPanel'
 import { RagPanel } from './panels/RagPanel'
 import { StructuredPanel } from './panels/StructuredPanel'
@@ -156,6 +158,8 @@ function App() {
         {sample === 'hyde' && <HydePanel provider={provider} />}
         {sample === 'semanticChunk' && <SemanticChunkPanel provider={provider} />}
         {sample === 'parentChild' && <ParentChildPanel provider={provider} />}
+        {sample === 'memoryExtract' && <MemoryExtractPanel provider={provider} />}
+        {sample === 'memoryCompare' && <MemoryComparePanel provider={provider} />}
       </AppShell.Main>
     </AppShell>
   )